requirements_tf.txt
bleach==1.5.0
certifi==2019.3.9
cloudpickle==0.8.1
cycler==0.10.0
decorator==4.4.0
dominate==2.3.5
enum34==1.1.6
html5lib==0.9999999
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
matplotlib==3.0.3
networkx==2.2
numpy==1.16.2
pandas==0.24.2
Pillow==5.4.1
protobuf==3.7.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0
toolz==0.9.0
tqdm==4.31.1
Werkzeug==0.15.1
